refactor(pfb): open print settings from the page menu
